refactor(video-conf): fewer components for the same conference UI
`ConferenceRoom` and `ConferenceThread` were the same component with a
different body: both opened the room, forced the embedded layout so the room UI
fits a narrow panel, and drew the same loading, success and error states — down
to an identical `NotSubscribedToRoomError` branch that would have drifted the
first time one of them was fixed. Now one `ConferenceRoomPanel` takes an
optional `tmid`, and `ConferenceChat` no longer carries two mutually exclusive
blocks to choose between them.

`ConferenceMemberRow` was twenty lines and one caller. It only needed its own
file because it called `useUserDisplayName`, which can't be called from a `map`
— but that hook is a setting read and a pure function, so the modal reads the
setting once and names each member itself.

The six components that were a single file in a folder of their own are now a
single file: a folder earns itself when there is a spec or a story to keep
beside the component.
