+++ apps/meteor/client/views/conference/ConferenceChat.spec.tsx
@@ -10,8 +10,7 @@ jest.mock('./ConferenceStoresReady', () => ({
 	__esModule: true,
 	default: ({ children }: { children: React.ReactNode }) => <div data-testid='chat-room'>{children}</div>,
 }));
-jest.mock('./ConferenceRoom', () => ({ __esModule: true, default: () => null }));
-jest.mock('./ConferenceThread', () => ({ __esModule: true, default: () => null }));
+jest.mock('./ConferenceRoomPanel', () => ({ __esModule: true, default: () => null }));
 
 // `withJohnDoe` fixes the logged-in id, so the member without access has to be that same user.
 const uid = 'john.doe';
