+++ apps/meteor/client/views/conference/ConferenceEmbeddedPage.tsx
@@ -14,11 +14,11 @@ import ConferenceThreadModal from './ConferenceThreadModal';
 import ConferenceUnauthorizedPage from './ConferenceUnauthorizedPage';
 import PageLoading from '../root/PageLoading';
 import CallMembersPanel from './components/CallMembersPanel/CallMembersPanel';
-import CallPanel from './components/CallPanel/CallPanel';
+import CallPanel from './components/CallPanel';
 import CallTimer from './components/CallTimer/CallTimer';
-import CallTopBar from './components/CallTopBar/CallTopBar';
+import CallTopBar from './components/CallTopBar';
 import ChatAccessNotice from './components/ChatAccessNotice/ChatAccessNotice';
-import ConferenceIframe from './components/ConferenceIframe/ConferenceIframe';
+import ConferenceIframe from './components/ConferenceIframe';
 import { useConferenceEmbedded } from './hooks/useConferenceEmbedded';
 import { useConferencePresenceLease } from './hooks/useConferencePresenceLease';
 import { useConferenceSubscription } from './hooks/useConferenceSubscription';
