+++ apps/meteor/client/views/conference/components/CallPanel.tsx
