+++ apps/meteor/client/views/conference/components/CallPanelHeader.tsx
